inventory service with aysnchronous communication to other services

+++ Ecom.OrderService/Entities/OrderItem.cs
@@ -4,10 +4,10 @@ namespace Ecom.OrderService.Entities
 {
     public class OrderItem : IEntity
     {
-        public Guid Id { get ; set; }
+        public Guid Id { get; set; }
         public Guid OrderId { get; set; }
         public Guid CatalogItemId { get; set; }
-        public double Price { get; set; }
+        public decimal Price { get; set; }
         public int Quantity { get; set; }
     }
 }
